+++ src/providers/tripUpdates/GtfsRtProvider.ts
@@ -1,6 +1,6 @@
 import axios, { AxiosRequestConfig } from 'axios';
 import config from '../../config/config';
-import { GtfsRTFeedProcessorSettings, GtfsRTFeedProcessor } from '../../lib/GtfsRTFeedProcessor';
+import { GtfsRTFeedProcessorSettings, GtfsRTFeedProcessor } from './lib/GtfsRTFeedProcessor';
 import { DataProvider } from '.';
 
 /**
